Simplify weather logic
Now everytime the user taps the update button, it will try to
update the weather, no matter if it kills battery. The user
is responsible enough of its usage.

It simplifies the code a lot.

It also makes more sense. The user at least see a visual feedback
with the location icon showing up.